pixel-main/lambda/requirements.txt
google-genai>=1.0
httpx>=0.27
Pillow>=10.0
boto3>=1.34
python-dotenv>=1.0
